+++ src/hooks/useSignup.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import React, { useState } from "react";
 import useInput from "@/hooks/useInput";
 import {
   authSendEmail,
@@ -6,6 +6,24 @@ import {
   authCheckEmailDuplicate,
   authSignUp,
 } from "@/api/auth";
+
+type InputProps = {
+  label: string;
+  value?: string;
+  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  type?: string;
+  message?: string;
+  placeholder?: string;
+  readOnly?: boolean;
+  isCertification?: {
+    title?: string;
+    size?: "large" | "medium" | "small";
+    disabled?: boolean;
+    theme?: string;
+    onClick?: (e?: React.MouseEvent<HTMLButtonElement>) => void;
+  };
+  inValid?: boolean;
+};
 
 export default function useSignup() {
   const [username, onChangeUserName] = useInput("");
@@ -18,7 +36,6 @@ export default function useSignup() {
   const [showText, setShowText] = useState(false);
   const [emailMessage, setEmailMessage] = useState("");
   const [emailButton, setEmailButton] = useState("인증");
-  const [emailValid, setEmailValid] = useState(false);
   const [duplicateCheck, setDuplicateCheck] = useState(false);
   const [codeMessage, setCodeMessage] = useState("");
   const [privateChecked, setPrivateChecked] = useState(false);
@@ -96,38 +113,90 @@ export default function useSignup() {
     setPrivateChecked(!privateChecked);
   };
 
+  const inputData: InputProps[] = [
+    {
+      label: "이메일주소",
+      placeholder: "예) rastle@rastle.com",
+      // buttonTitle: "전송",
+      onChange: onChangeEmailHandler,
+      message:
+        email.length > 0 && !isValidEmail(email)
+          ? "이메일 형식이 틀렸습니다"
+          : emailMessage,
+      inValid: email.length > 0 && (!isValidEmail(email) || duplicateCheck),
+      isCertification: {
+        title: emailButton,
+        disabled: (email.length > 0 && !isValidEmail(email)) || codeMatch,
+        onClick: () => {
+          sendEmailCode();
+        },
+      },
+    },
+    {
+      label: "인증번호",
+      placeholder: "",
+      onChange: onChangeCode,
+      message: codeMessage,
+      inValid: codeMessage === "코드가 일치하지 않습니다" && !codeMatch,
+      isCertification: {
+        title: "확인",
+        disabled: codeMatch,
+        onClick: () => {
+          checkEmailCode();
+        },
+      },
+    },
+    {
+      label: "비밀번호",
+      type: "password",
+      placeholder: "영문,숫자,특수문자 조합 8~16자",
+      onChange: onChangePassword,
+      message:
+        password.length > 0 && password.length < 8
+          ? "비밀번호를 8자리 이상 입력하세요 "
+          : "",
+      inValid: password.length > 0 && password.length < 8,
+    },
+    {
+      label: "비밀번호 확인",
+      type: "password",
+      onChange: onChangePasswordCheck,
+      message:
+        password !== passwordCheck && passwordCheck.length > 0
+          ? "비밀번호가 일치하지 않습니다."
+          : "",
+      inValid: password !== passwordCheck && passwordCheck.length > 0,
+    },
+    {
+      label: "이름",
+      placeholder: "예) 홍레슬",
+      onChange: onChangeUserName,
+    },
+    {
+      label: "휴대폰 번호",
+      placeholder: "예) 01012345678",
+      onChange: onChangePhoneNumber,
+      inValid:
+        phoneNumber.length > 0 &&
+        !/^(01[016789]{1})[0-9]{3,4}[0-9]{4}$/.test(phoneNumber),
+      message:
+        phoneNumber.length > 0 &&
+        !/^(01[016789]{1})[0-9]{3,4}[0-9]{4}$/.test(phoneNumber)
+          ? "유효하지 않은 전화번호입니다"
+          : "",
+    },
+  ];
+
   return {
     username,
-    onChangeUserName,
-    onChangeEmailHandler,
-    email,
-    onChangeEmail,
-    code,
-    onChangeCode,
     password,
-    onChangePassword,
     passwordCheck,
-    onChangePasswordCheck,
     phoneNumber,
-    onChangePhoneNumber,
-    sendEmailCode,
     showText,
-    setShowText,
     codeMatch,
-    setCodeMatch,
-    checkEmailCode,
-    checkEmailDuplicated,
-    isValidEmail,
-    emailMessage,
-    setEmailMessage,
-    emailButton,
-    setEmailButton,
-    emailValid,
-    setEmailValid,
-    duplicateCheck,
-    codeMessage,
     signUp,
     togglePrivate,
     privateChecked,
+    inputData,
   };
 }
